+++ src/lib.rs
@@ -10,11 +10,11 @@
 
 #![doc = include_str!("../README.md")]
 
-// Make available for public use.
 mod circuit;
 mod complex;
 mod error;
 
+// Make available for public use.
 pub use circuit::printer::Printer;
 pub use circuit::states;
 pub use circuit::{Circuit, Gate, Measurement};
